+++ crates/runtime/src/lib.rs
@@ -35,6 +35,7 @@ use app::App;
 use cache::QueryResultsCacheProvider;
 use component::dataset::{self, Dataset};
 use config::Config;
+use datafusion::query::query_history;
 use datafusion::SPICE_RUNTIME_SCHEMA;
 use futures::future::join_all;
 use futures::StreamExt;
@@ -69,8 +70,6 @@ pub mod object_store_registry;
 pub mod objectstore;
 mod opentelemetry;
 pub mod podswatcher;
-pub mod query_context;
-pub mod query_history;
 pub mod spice_metrics;
 pub mod status;
 pub mod timing;
